pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clip_finetune"
version = "1.0.0"
description = "Fine-tuning a CLIP-like model for monthly image data."
authors = [
    { name = "Saman Nia", email = "saman.p.nia@gmail.com" }
]
readme = "README.md"

dependencies = [
    "torch",
    "pillow",
    "open_clip_torch",
    "pyyaml",
    "boto3",  # boto3 for real S3 usage
    "mlflow", # for AWS monitoring usage
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]